% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{aggregate_and_filter}
\alias{aggregate_and_filter}
\title{Aggregates and filters outputs of a single rTREES run}
\usage{
aggregate_and_filter(
  rtrees_results,
  root_shoot_soil,
  interval_to_return = 48,
  vars_to_return,
  aggregate_time_steps = FALSE
)
}
\arguments{
\item{rtrees_results}{Results from rTREES::Clean_rTREES_output}

\item{root_shoot_soil}{see rTREES_parallel_run}

\item{interval_to_return}{see rTREES_parallel_run}

\item{vars_to_return}{see rTREES_parallel_run}

\item{aggregate_time_steps}{see rTREES_parallel_run}
}
\value{
Aggregated and filters rTREES results
}
\description{
Aggregates and filters outputs of a single rTREES run
}
