% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{attempt_driver_repair}
\alias{attempt_driver_repair}
\title{Attempt Driver Repair}
\usage{
attempt_driver_repair(na_df)
}
\arguments{
\item{na_df}{Data frame of driver containing NA values. Must contain all columns properly named. See BuildDriver() or Driver_inputs()}
}
\value{
Data frame of driver with NA values filled or imputed as well as possible.
}
\description{
Attempt Driver Repair
}
\examples{
bdr<-BuildDriver()
bdr[matrix(
  sample(
    x = c(TRUE,FALSE),
    size = nrow(bdr)*ncol(bdr),
    prob = c(0.01,0.99),
    replace = TRUE
  ),
  nrow = nrow(bdr),
  ncol = ncol(bdr)
)]<- NA # Add random NA to simulate a driver with missing values.
demo_repaired_driver<-attempt_driver_repair(bdr)
Visual_driver(demo_repaired_driver)
}
