% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{attempt_p_repair}
\alias{attempt_p_repair}
\title{Repair Parameters}
\usage{
attempt_p_repair(params_in)
}
\arguments{
\item{params_in}{Data frame for broken TREES parameters.}
}
\value{
list containing attempted repairs and error count.
}
\description{
Repair Parameters
}
\examples{
demo_repair_p<-attempt_p_repair(
 params_in = ExampleParameters()[
 sample(1:174,170,replace = FALSE),
 ]#example of missing rows in an incorrect order
)
}
