% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{BuildDriver}
\alias{BuildDriver}
\title{Build Driver}
\usage{
BuildDriver(
  lat = 35.818,
  long = -106.305,
  alt = 2150,
  ndays = 365,
  year = 2023,
  start_jday = 1,
  dewpoint = 12,
  co2_atm = 410.5,
  wind_speed = 1.01,
  precipdays = c(5, 30, 50, 80, 90, 110, 140, 150, 200, 210, 290),
  precip_amount = 6.25,
  temp_scalar = 1
)
}
\arguments{
\item{lat}{Latitude}

\item{long}{Longitude}

\item{alt}{Altitude}

\item{ndays}{Number of days to run the driver}

\item{year}{Year}

\item{start_jday}{=1,}

\item{dewpoint}{= 12,}

\item{co2_atm}{=410.5,}

\item{wind_speed}{=1.01,}

\item{precipdays}{= c(20, 30, 50, 80, 90, 110, 140, 150, 200, 210, 290)}

\item{precip_amount}{=6.25 how much precip occurs in a precip day}

\item{temp_scalar}{Scales the tempertature vectors}
}
\value{
Data frame of a valid driver
}
\description{
Build Driver
}
\examples{
demo_driver<-BuildDriver(
  lat = 35.818,
  long = -106.305,
  alt = 2150.0,
  ndays = 50,
  year = 2023,
  start_jday=1,
  dewpoint = 12,
  co2_atm=410.5,
  wind_speed=1.01,
  precipdays = c(20, 30, 50, 80, 90, 110, 140, 150, 200, 210, 290),
  precip_amount = 6.25,
  temp_scalar=1
)
}
