% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cap_first_letter}
\alias{cap_first_letter}
\title{Captize the first letter of a word.}
\usage{
cap_first_letter(x)
}
\arguments{
\item{x}{A word.}
}
\value{
The same word with the first letter captized.
}
\description{
Captize the first letter of a word.
}
