% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{ChangeParameter}
\alias{ChangeParameter}
\title{Change parameters}
\usage{
ChangeParameter(
  base_parameters,
  NewValues = data.frame(SLA = 0),
  VERBOSE = TRUE
)
}
\arguments{
\item{base_parameters}{Data.frame of the initial parameter values in first column or or path to text file of parameters values (one on each line in correct order).}

\item{NewValues}{Data.frame with columns named for the parameters and one row containg the new values for those parameters. Verify correct names by checking ExampleParameters() or param_search()
This only works with one row of values at a time.}

\item{VERBOSE}{Set verbosity to TRUE or FALSE}
}
\value{
Data frame of the base parameters with all modifications from NewValues and validated against default parameter set.
}
\description{
Change parameters
}
\examples{
demo_new_parameters<-ChangeParameter(
   base_parameters=ExampleParameters(),
   NewValues=data.frame(
     "initial_conductivity_root"=25,
     "Gsref0"=0.1
   ),
   VERBOSE=TRUE
)
#'
}
