% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{Clean_rTREES_output}
\alias{Clean_rTREES_output}
\title{Clean rTREES output}
\usage{
Clean_rTREES_output(rTREES_output = list())
}
\arguments{
\item{rTREES_output}{raw rTREES list output from a single simulation.}
}
\value{
A list of 3 data frames corresponding to the original TREES outputs of sim,hyd, and leaf.
}
\description{
Clean rTREES output
}
\examples{
demo_standard_run<-rTREES(
  env_driver = BuildDriver(ndays = 10),
  base_parameters = ExampleParameters(),
  root_shoot_soil = ExampleRSS()
)
demo_cleaned_results<-Clean_rTREES_output(
  rTREES_output=demo_standard_run

)
}
