% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{CleanRootShootParams}
\alias{CleanRootShootParams}
\title{Load, prepare, and validate root/shoot module parameters.}
\usage{
CleanRootShootParams(root_shoot_soil)
}
\arguments{
\item{root_shoot_soil}{A path to a root_shoot_soil file or a data.frame() first column is parameter names and the second column contains values.}
}
\value{
list() containing root/shoot module parameters properly formatted for rTREES
}
\description{
Load, prepare, and validate root/shoot module parameters.
}
\examples{
demo_rss_from_file<-CleanRootShootParams(
 root_shoot_soil = system.file("examples","demo_root_shoot_soil",package="rTREES")
)
}
