% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{cosine_sim}
\alias{cosine_sim}
\title{Cosine Similarity ranks the similarity of text.}
\usage{
cosine_sim(text1, text2, token_length = 3, relative_corpus_table = NA)
}
\arguments{
\item{text1}{first string or vector of strings to compare}

\item{text2}{second}

\item{token_length}{Optional size if token chunks}

\item{relative_corpus_table}{Optional table of all possible values for weighing uniqueness of tokens}
}
\value{
cosine similarity of text1 and text2
}
\description{
Cosine Similarity ranks the similarity of text.
}
\examples{
cosine_sim(
 text1="the root carbon",
 text2="root nitrogen",
 token_length=6,
 relative_corpus_table=table(tiny_tokenizer("A collection of text to represent the expected
  distribution of tokens before evaluating
  the text of interest to avoid over weighing
   common tokens such as articles."))
)
}
