% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{depth}
\alias{depth}
\title{Get the max depth of a list.
Adopted from https://stackoverflow.com/questions/13432863/determine-level-of-nesting-in-r#comment18364266_13433689}
\usage{
depth(this)
}
\arguments{
\item{this}{the list of interest}
}
\value{
Maximum depth of list as an integer.
}
\description{
Get the max depth of a list.
Adopted from https://stackoverflow.com/questions/13432863/determine-level-of-nesting-in-r#comment18364266_13433689
}
\examples{
depth(ExampleRSS())
}
