% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_trees.R
\name{download_fish_full_tree}
\alias{download_fish_full_tree}
\title{Get full fish phylogeny provided by the \href{https://fishtreeoflife.org/about/}{Fish Tree of Life}}
\usage{
download_fish_full_tree(path_save_to = NULL, read_into_R = TRUE)
}
\arguments{
\item{path_save_to}{Where do you want to save the file? The default is \code{NULL},
which means that not to save it locally. When specify, specify to the folder/directory.}

\item{read_into_R}{Do you want to read the loaded phylogenies into R? Default is \code{TRUE}.}
}
\description{
Get the 100 phylogenies, each with 31516 tips.
The \verb{Complete “all-taxon assembled” chronograms} option from
its \href{https://fishtreeoflife.org/downloads/}{downloads page}.
}
