% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{Driver_inputs}
\alias{Driver_inputs}
\title{Driver_inputs}
\usage{
Driver_inputs()
}
\value{
Data frame of exact parameter names and their descriptions for drivers.
}
\description{
Driver_inputs
}
