% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{ExampleRSS}
\alias{ExampleRSS}
\title{Example Root, Shoot, and Soil layer/module parameters}
\usage{
ExampleRSS(file = FALSE)
}
\arguments{
\item{file}{If TRUE then a character string that can be saved to a text file will be produced.}
}
\value{
List of valid Root, Shoot, and Soil layer/module parameters
}
\description{
Example Root, Shoot, and Soil layer/module parameters
}
