% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{ez_dt}
\alias{ez_dt}
\title{Simplifies cleaning up TREES datetime output.}
\usage{
ez_dt(year, jday, hour, min, tz = "+0000")
}
\arguments{
\item{year}{year}

\item{jday}{day of year (julian day)}

\item{hour}{hour (24 hour time)}

\item{min}{minutes (all non 0 values are assumed 30 min)}

\item{tz}{timezone offset in the format of two digit hour and two digit min preceded by + or -}
}
\value{
datetime POSIXct object
}
\description{
Simplifies cleaning up TREES datetime output.
}
