% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_graft_status}
\alias{get_graft_status}
\title{Extract grafting status information as a data frame}
\usage{
get_graft_status(tree)
}
\arguments{
\item{tree}{A phylogeny generated by \code{get_tree(...)} with trailing stars in tip labels.}
}
\value{
A tibble with three columns: tip_label, species, and status.
}
\description{
Extract grafting status information as a data frame
}
