% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tree.R
\name{get_tree}
\alias{get_tree}
\title{Get one or multiple trees from megatree(s)}
\usage{
get_tree(
  sp_list,
  tree,
  taxon = NULL,
  scenario = c("at_basal_node", "random_below_basal"),
  show_grafted = FALSE,
  tree_by_user = FALSE,
  mc_cores = future::availableCores() - 2,
  .progress = "text",
  fish_tree = c("timetree", "all-taxon"),
  mammal_tree = c("vertlife", "phylacine"),
  bee_tree = c("maximum-likelihood", "bootstrap"),
  plant_tree = c("tree_plant_otl", "tree_plant_Carruthers", "tree_plant_n100_Carruthers"),
  dt = TRUE
)
}
\arguments{
\item{sp_list}{A character vector or a data frame with at least three columns: species, genus, family. Species column
holds the species for which we want to have a phylogeny. It can also have two optional columns:
close_sp and close_genus. We can specify the closest species/genus of the species based on
expert knowledge. If specified, the new species will be grafted to that particular location.

It can also be a string vector if \code{taxon} is specified. Though it probably is a better idea
to prepare your data frame with \code{\link[=sp_list_df]{sp_list_df()}}. The string vector can also have the same format
as that required by Phylomatic (i.e., family/genus/genus_sp).}

\item{tree}{A mega-tree with class \code{phylo} or a list of mega-trees with class \code{multiPhylo}.
Optional if \code{taxon} is specified, in which case, a default
mega-phylogeny (or a set of 100 randomly selected posterior phylogenies) will be used
(see their own documentations from the \code{megatrees} package).
\itemize{
\item For amphibian, the mega-trees are loaded via \code{\link[megatrees:get_tree_amphibian_n100]{megatrees::get_tree_amphibian_n100()}}.
\item For bee, the mega-tree is \link[megatrees:tree_bee]{megatrees::tree_bee}, with the bootstrap option loaded via \code{\link[megatrees:get_tree_bee_n100]{megatrees::get_tree_bee_n100()}}.
\item For butterfly, the mega-tee is \link[megatrees:tree_butterfly]{megatrees::tree_butterfly}.
\item For bird, the mega-trees are loaded via \code{\link[megatrees:get_tree_bird_n100]{megatrees::get_tree_bird_n100()}}.
\item For fish, the mega-tree is \link[megatrees:tree_fish_12k]{megatrees::tree_fish_12k}, with the all-taxon option loaded via \code{\link[megatrees:get_tree_fish_32k_n50]{megatrees::get_tree_fish_32k_n50()}}.
\item For mammal, the default mega-trees are loaded via \code{\link[megatrees:get_tree_mammal_n100_vertlife]{megatrees::get_tree_mammal_n100_vertlife()}}, with \code{\link[megatrees:get_tree_mammal_n100_phylacine]{megatrees::get_tree_mammal_n100_phylacine()}} be the other option.
\item For plant, the default mega-tree is \link[megatrees:tree_plant_otl]{megatrees::tree_plant_otl} (Smith and Brown 2018);
\link[megatrees:tree_plant_Carruthers]{megatrees::tree_plant_Carruthers} (Carruthers et al. 2026) is also available; 100 posterior trees are loaded via
\code{\link[megatrees:get_tree_plant_n100_Carruthers]{megatrees::get_tree_plant_n100_Carruthers()}}.
\item For reptile, the mega-trees are loaded via \code{\link[megatrees:get_tree_reptile_n100]{megatrees::get_tree_reptile_n100()}}.
\item For shark, ray, and chimaeras, the mega-trees are loaded via \code{\link[megatrees:get_tree_shark_ray_n100]{megatrees::get_tree_shark_ray_n100()}}.
}}

\item{taxon}{The taxon of species in the \code{sp_list}. Currently, can be \code{amphibian}, \code{bird}, \code{fish}, \code{mammal}, \code{plant}, \code{reptile}, or \code{shark_ray}.}

\item{scenario}{How to insert a species into the mega-tree?
\itemize{
\item In both scenarioes, if there is only 1 species in the genus or family, a new node will be inserted to
the middle point of this only species' branch length and the new species will be attached to this new
node.
\item If \code{scenario = "at_basal_node"}, a species is attached to the basal node of the same genus or the same family
if the mega-tree does not have any species of this genus.
\item If \code{scenario = "random_below_basal"}, a species is attached to a randomly selected node that is at or below the
basal node of the same genus of the same family if the mega-tree does not have any species in this genus.
The probability of node been selected is proportional to its branch length.
Because of the random sampling involved, you may want to run several times to get a collection of
derived phylogenies.
}}

\item{show_grafted}{Whether to indicate which species was grafted onto the mega-tree.
If \code{TRUE}, a \code{*} will be appended to the species name on the tip if it was grafted within
the same genus; \verb{**} will be appended if it was grafted within the same family.}

\item{tree_by_user}{Is the mega-tree provided by user? Default is \code{FALSE} but it will be automatically set to \code{TRUE} when the class of \code{tree} is
\code{multiPhylo} since we don't provide any such mega-trees here.}

\item{mc_cores}{Number of cores to parallel processing when \code{tree} is a list of large number of trees. The default is the number of available cores minus 2.}

\item{.progress}{Form of progress bar, default to be text.}

\item{fish_tree}{Which fish tree do you want to use? If it is "timetree" (default), it will be the smaller time tree with 11638 species
that all have sequence data; if it is "all-taxon", then it will be the 100 larger posterior phylogenies with 31516 soecues.}

\item{mammal_tree}{Which set of mammal trees to use? If it is "vertlife" (default), then 100 randomly selected posterior phylogenies provided
by Vertlife will be used; if it is "phylacine", then 100 randomly selected posterior phylogenies provided by PHYLACINE will be used.}

\item{bee_tree}{Which bee tree to use? If it is "maximum-likelihood" (default), the a single maximum likelihood tree will be used. If it is "bootstrap", then a set of 100 randomly selected posterior phylogenies will be used. All trees are provided by the \href{http://beetreeoflife.org}{Bee Tree of Life}.}

\item{plant_tree}{Which plant tree to use? If \code{"tree_plant_otl"} (default), the Smith and Brown (2018) tree
is used. If \code{"tree_plant_Carruthers"}, the single best-scoring tree from Carruthers et al. (2026) is used.
If \code{"tree_plant_n100_Carruthers"}, 100 randomly selected posterior phylogenies from Carruthers et al. (2026)
are downloaded and used (large file, ~135 MB, cached after first use).}

\item{dt}{Whether to use data.table version to bind tips \link{bind_tip}. The default is \code{TRUE} as it maybe slightly faster.}
}
\value{
A phylogeny for the species required, with class \code{phylo};
or a list of phylogenies with class \code{multiPhylo} depends on the input \code{tree}. Within each phylogeny, the grafted status of all species was saved as a data frame named as "graft_status".
}
\description{
For some taxa groups, there are multiple posterior megatrees. It is a common
task to derive a phylogeny from each of these (or a random subset of) megatrees.
}
\details{
Derive a phylogeny from a mega-tree

For a list of species, generate a phylogeny or multiple phylogenies from a
provided mega-tree or mega-trees. If a species is
not in the mega-tree, it will be grafted to the mega-tree with two scenarios.
}
\examples{
\donttest{
test_sp <- c(
  "Serrasalmus_geryi", "Careproctus_reinhardti", "Gobiomorphus_coxii",
  "Periophthalmus_barbarus", "Prognichthys_glaphyrae", "Barathronus_bicolor",
  "Knipowitschia_croatica", "Rhamphochromis_lucius", "Neolissochilus_tweediei",
  "Haplochromis_nyanzae", "Astronesthes_micropogon", "Sanopus_reticulatus"
)
test_tree <- get_tree(
  sp_list = test_sp,
  taxon = "fish",
  show_grafted = TRUE
)
}
}
