% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Custom operator to more cleanly perform !\%in\%.}
\usage{
a \%notin\% b
}
\arguments{
\item{a}{The left side vector of elements to look for in the right side vector.}

\item{b}{The right side vector of elements to look through for matches to the left side vector.}
}
\value{
TRUE or FALSE a is not in b
}
\description{
Custom operator to more cleanly perform !\%in\%.
}
\examples{
c("bananas")\%notin\%c("apple","cat","music")
c("apple")\%notin\%c("apple","cat","music")
}
