% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{lnorm_to_norm}
\alias{lnorm_to_norm}
\title{For getting mean and sd from lognormal parameters}
\usage{
lnorm_to_norm(lm, ls)
}
\arguments{
\item{lm}{meanlog or location parameter}

\item{ls}{Standard deviation of desired distribution}
}
\value{
A named vector of doubles for the mean and sd.
}
\description{
For getting mean and sd from lognormal parameters
}
