% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{Load_TREES_files}
\alias{Load_TREES_files}
\title{Read in original style TREES input files}
\usage{
Load_TREES_files(driver = NA, parameters = NA, root_shoot_soil = NA)
}
\arguments{
\item{driver}{path to tab deliminated driver file (typically ends in .txt)}

\item{parameters}{path to parameter file (typically ends in .p)}

\item{root_shoot_soil}{path to root/shoot module file (typically named root_shoot_soil previously called param_mod)}
}
\description{
Read in original style TREES input files
}
\examples{
demo_inputs_from_file<-Load_TREES_files(
  driver=system.file("examples","demo_driver.txt",package="rTREES"),#"path/to/driver.txt",
  parameters = system.file("examples","demo_parameter.p",package="rTREES"),#"path/to/parameters.p",
 root_shoot_soil = system.file("examples","demo_root_shoot_soil",package="rTREES")#"path/to/rss"
)
}
