% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{long_df_hyd_parallel}
\alias{long_df_hyd_parallel}
\title{Clean and make long the hyd data from a parallel run}
\usage{
long_df_hyd_parallel(par_res)
}
\arguments{
\item{par_res}{The output of a parallel rTREES run.}
}
\value{
A data.frame() of leaf data in long format.
}
\description{
Clean and make long the hyd data from a parallel run
}
