% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{norm_to_lnorm}
\alias{norm_to_lnorm}
\title{For getting lognormal parameters from a desired mean and sd}
\usage{
norm_to_lnorm(m, s)
}
\arguments{
\item{m}{Mean of desired distribution}

\item{s}{Standard deviation of desired distribution}
}
\value{
A named vector of doubles for the meanlog and sdlog.
}
\description{
For getting lognormal parameters from a desired mean and sd
}
