% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{number_of_cores}
\alias{number_of_cores}
\title{This is a very simple function for portabl core detection.}
\usage{
number_of_cores()
}
\value{
The number of cores detected by your system
}
\description{
This is a very simple function for portabl core detection.
}
