% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{param_idx}
\alias{param_idx}
\title{Get the index of a parameter
Used mostly internally to rTREES.}
\usage{
param_idx(parameter_name = "")
}
\arguments{
\item{parameter_name}{valid parameter name as is found in ExampleParameters()}
}
\value{
returns the index of the input parameter name.
}
\description{
Get the index of a parameter
Used mostly internally to rTREES.
}
