% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{param_search}
\alias{param_search}
\title{Search for the exact name of a parameter when you are unsure
Can't recall the excat name of a parameter? Happens to me all the time. Here is a little function I tossed together to **sometimes** help with that. The logic is the same as that in the attempt_p_repair function, so don't expect perfection.}
\usage{
param_search(query = "", n_return = 5)
}
\arguments{
\item{query}{key words or guesses of the parameters name}

\item{n_return}{how many results would you liked returned? Defaults to top 5}
}
\value{
The contents of the ExampleParameter() row, including the parameter name, that best matches the query.
}
\description{
Search for the exact name of a parameter when you are unsure
Can't recall the excat name of a parameter? Happens to me all the time. Here is a little function I tossed together to **sometimes** help with that. The logic is the same as that in the attempt_p_repair function, so don't expect perfection.
}
