% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{Possible_Outputs}
\alias{Possible_Outputs}
\title{Possible variables (Need to add explainations to these)}
\usage{
Possible_Outputs(n_layers = 4)
}
\arguments{
\item{n_layers}{The number of soil layers}
}
\value{
A vector of all possible rTREES variable outputs
}
\description{
Possible variables (Need to add explainations to these)
}
