% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{prep_rTREES_parallel}
\alias{prep_rTREES_parallel}
\title{Prepare inputs for spooky_parallel}
\usage{
prep_rTREES_parallel(
  base_parameters = parameters,
  new_parameters,
  drivers = NA,
  rss = NA,
  vars_to_return = "all",
  interval_to_return = 12,
  aggregate_time_steps = FALSE,
  verbosity = FALSE,
  softerror_max = 5
)
}
\arguments{
\item{base_parameters}{Base set of input parameters}

\item{new_parameters}{New parameters to test}

\item{drivers}{list of drivers to run}

\item{rss}{Root shoot and soil parameters (list of list)}

\item{vars_to_return}{Which variables are you interested in?}

\item{interval_to_return}{Time interval to be returned. 0.5 is half hourly, 1 is hourly, 12 is twice daily, 24 is daily, et}

\item{aggregate_time_steps}{Should the time steps not included be averaged to the interval returned?}

\item{verbosity}{Include standard output?}

\item{softerror_max}{How many soft errors should be tolerated? It is normal for long runs, especially over seasons, to have a few soft errors without it causing any issues.}
}
\value{
args list for spooky_parallel
}
\description{
Prepare inputs for spooky_parallel
}
