% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{ReformRawP}
\alias{ReformRawP}
\title{Clean parameter values}
\usage{
ReformRawP(dirty_vector = c("2013/tY"))
}
\arguments{
\item{dirty_vector}{The vector of character strings to be cleaned}
}
\value{
Splits the parameter values, in a character vector from reading a file, from the rest of the contents(notes) and returns them as a dataframe.
}
\description{
Clean parameter values
}
\examples{
demo_raw_p<-ReformRawP(
dirty_vector =readLines(
  con = system.file("examples","demo_parameter.p",package="rTREES")
  )
)
}
