% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{retry_spooky}
\alias{retry_spooky}
\title{If any of the elements of spooky_parallel failed to complete (only visible if debug was set to TRUE), then you can use this function to retry running only the ones that failed. You can use this to also set longer run times (actual maximum run time, not a scalar) or turn on verbosity for viewing job_stdout.}
\usage{
retry_spooky(results_debug, n_cores, max_time = 0, verbosity = FALSE)
}
\arguments{
\item{results_debug}{Secong list element output from rTREES_parallel_run() or rTREES_par_finsh() when debug is TRUE OR output from rTREES::spooky_parallel.}

\item{n_cores}{Number of cores to use}

\item{max_time}{Actual maximum run time}

\item{verbosity}{Use verbosity (only applies if verbosity is an argument of the function being run)}
}
\value{
Debug results with new results of runs that didn't complete before. NOTE this does not guarantee the run will work. On a second failure from timeout, try longer max_time. If it is any other failure, investigate stdout or other debug information.
}
\description{
If any of the elements of spooky_parallel failed to complete (only visible if debug was set to TRUE), then you can use this function to retry running only the ones that failed. You can use this to also set longer run times (actual maximum run time, not a scalar) or turn on verbosity for viewing job_stdout.
}
