% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rm_stars}
\alias{rm_stars}
\title{Remove trailing *}
\usage{
rm_stars(tree)
}
\arguments{
\item{tree}{A phylogeny generated by \code{get_tree(..., show_grafted = TRUE)} with trailing stars in tip labels.}
}
\value{
A phylogeny after removing trailing stars.
}
\description{
Remove trailing *
}
