% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rTREES}
\alias{rTREES}
\title{This function prepares the input files and runs simulator. Only the first three parameters are mandatory.}
\usage{
rTREES(
  env_driver,
  base_parameters,
  root_shoot_soil,
  which_out = as.logical(c(1, 1, 1)),
  verbosity = FALSE,
  softerror_max = 5L
)
}
\arguments{
\item{env_driver}{data.frame() containing environmental data for each time step. See Driver_inputs() for explanations of required columns. Use ValidateDriver() to check for errors.}

\item{base_parameters}{data.frame() where the first column is parameter names and the second column contains values. See Possible_parameters() for explanations of required parameters.}

\item{root_shoot_soil}{list() where each item is a vector for each type of parameter in orig param_mod file}

\item{which_out}{vector() of length 3 of boolean values to restrict which outputs are returned. Format is in order sim,hyd,leaf}

\item{verbosity}{true displays extra verbosity.}

\item{softerror_max}{integer for how sensitive TREES should be before failing. This mostly exists to allow infrequent hydraulic errors to only stop the simulation if they occur a certain number of times.}
}
\value{
a list of leaf areas, simulation outputs, and hydraulic outputs
}
\description{
This function prepares the input files and runs simulator. Only the first three parameters are mandatory.
}
\examples{
demo_standard_run<<-rTREES(
     env_driver = BuildDriver(ndays = 10),
     base_parameters = ExampleParameters(),
     root_shoot_soil = ExampleRSS()
 )
demo_cleaned_run<<-Clean_rTREES_output(
    rTREES_output = demo_standard_run
    )
}
