% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{rTREES_calib_function}
\alias{rTREES_calib_function}
\title{Timeout calibration function for spooky_parallel when using rTREES}
\usage{
rTREES_calib_function(list_args_list, all_pkgs, run_func, n_cores)
}
\arguments{
\item{list_args_list}{Outputs from prep_rTREES_parallel()}

\item{all_pkgs}{Required packages}

\item{run_func}{Function to run (rTREES_wrapper)}

\item{n_cores}{Number of cores to use}
}
\value{
time estimate for longest run out of args list
}
\description{
Timeout calibration function for spooky_parallel when using rTREES
}
