% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{rTREES_par_finish}
\alias{rTREES_par_finish}
\title{Clean up final outputs from spooky_parallel}
\usage{
rTREES_par_finish(job_ls, debug = FALSE)
}
\arguments{
\item{job_ls}{Output of spooky_parallel when using rTREES}

\item{debug}{When TRUE retain information from callr run. This includes, errors, exit conditions, and more. If rTREES verbosity is also TRUE, that will be stored in the job_stdout element. Use cat() when trying to view job_stdout.}
}
\value{
If debug is FALSE, returns the typical rTREES parallel results. If debug is TRUE, returns a list with typical rTREES parallel results and the full job outputs from callr.
}
\description{
Clean up final outputs from spooky_parallel
}
