% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{rTREES_spooky_wrapper}
\alias{rTREES_spooky_wrapper}
\title{rTREES wrapper for spooky_parallel. This exists because spooky_parallel only takes and passes one arg.}
\usage{
rTREES_spooky_wrapper(
  args_in = list(base_parameters = parameters, env_driver = NA, root_shoot_soil = NA,
    vars_to_return = "all", interval_to_return = 12, aggregate_time_steps = FALSE,
    verbosity = FALSE, softerror_max = 5)
)
}
\arguments{
\item{args_in}{named list of rTREES inputs}
}
\value{
cleaned and aggregated results of a single element from the args list.
}
\description{
rTREES wrapper for spooky_parallel. This exists because spooky_parallel only takes and passes one arg.
}
