% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{spooky_parallel}
\alias{spooky_parallel}
\title{Run parallel jobs
This is a generic function that will run any function over all provided argument combinations in parallel using callr.}
\usage{
spooky_parallel(
  list_args_list = list(job_1 = list(drivers = list(dr1 = rTREES::BuildDriver(ndays = 10,
    start_jday = 100)), base_parameters = ExampleParameters(), new_parameters =
    data.frame(), rss = ExampleRSS(), verbosity = FALSE, softerror_max = 5, )),
  all_pkgs = c("rTREES"),
  run_func = rTREES::rTREES,
  n_cores = rTREES::number_of_cores() - 2,
  timeout_scalar = 0,
  calibration_fun = NA,
  progressbar = TRUE
)
}
\arguments{
\item{list_args_list}{Arguments for each iteration of the function to run.}

\item{all_pkgs}{Packages required for function to run}

\item{run_func}{Function that will run arguments. Note, do not include () and try to include function package. Ex package::function}

\item{n_cores}{How many parallel cores to use}

\item{timeout_scalar}{If not zero and a calibration function is provided this value is multiplied by the estimated maximum time calculated by the calibration_fun to determine when a run should timeout.}

\item{calibration_fun}{A function that will determine the maximum time a single run should take. If only a number is provided it will be treated as the maximum time in seconds.}

\item{progressbar}{If TRUE a progressbar will be shown in the console. This fails in snippets.}
}
\value{
list of results where each element of the list corresponds to a row of parameters in new_parameters
}
\description{
Run parallel jobs
This is a generic function that will run any function over all provided argument combinations in parallel using callr.
}
