% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{taxa_supported}
\alias{taxa_supported}
\title{Taxonomic groups supported}
\format{
An object of class \code{character} of length 8.
}
\usage{
taxa_supported
}
\description{
Supported taxonomic groups with mega-trees provided in the {megatrees} package.
}
\keyword{datasets}
