% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{tiny_tokenizer}
\alias{tiny_tokenizer}
\title{Tokenizer}
\usage{
tiny_tokenizer(text_string, token_length = 3)
}
\arguments{
\item{text_string}{Character string or vector to be broken into tokens}

\item{token_length}{Optional size if token chunks}
}
\value{
Vector of tokens.
}
\description{
Tokenizer
}
\examples{
tiny_tokenizer(
  text_string = "Some example text",
  token_length = 3
)
}
