% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tree_bird_ericson}
\alias{tree_bird_ericson}
\title{Mega-tree of birds from Bird Tree.}
\format{
A phylogeny with class "phylo". It is also a list. Compare with a normal phylo object, it has another data frame \code{tree_bird_ericson$genus_family_root}, which provides the root nodes information for every unique genus and family in the phylogeny. Such information can be used to insert new tips onto the phylogeny later.
}
\source{
\url{https://birdtree.org/downloads/}
}
\usage{
tree_bird_ericson
}
\description{
This tree was the 288th \verb{set.seed(123); sample(1:1000, 1)} of those 10000 trees with Ericson backbond generated by Jetz et al. (2012). Node labels were added here. This tree will be the default one for bird.
}
\references{
Jetz, W., Thomas, G. H., Joy, J. B., Hartmann, K., & Mooers, A. O. (2012). The global diversity of birds in space and time. Nature, 491(7424), 444.
}
\seealso{
Other trees: 
\code{\link{tree_bird_hackett}},
\code{\link{tree_fish}},
\code{\link{tree_mammal}},
\code{\link{tree_plant_otl}}
}
\concept{trees}
\keyword{datasets}
