% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tree_fish}
\alias{tree_fish}
\title{Mega-tree of fish from The Fish Tree of Life}
\format{A phylogeny with class "phylo". It is also a list. Compare with a normal phylo object, it has another data frame \code{tree_fish$genus_family_root}, which provides the root nodes information for every unique genus and family in the phylogeny. Such information can be used to insert new tips onto the phylogeny later.}
\source{
\url{https://fishtreeoflife.org/downloads/actinopt_12k_treePL.tre.xz}
}
\usage{
tree_fish
}
\description{
This tree was generated by Rabosky et al. (2018). Node labels were added here.
}
\references{
Rabosky, D. L., Chang, J., Title, P. O., Cowman, P. F., Sallan, L., Friedman, M., ... & Alfaro, M. E. (2018). An inverse latitudinal gradient in speciation rate for marine fishes. Nature, 559(7714), 392.
}
\seealso{
Other trees: \code{\link{tree_bird_ericson}},
  \code{\link{tree_bird_hackett}},
  \code{\link{tree_mammal}}, \code{\link{tree_plant_otl}}
}
\concept{trees}
\keyword{datasets}
