% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tree_mammal}
\alias{tree_mammal}
\title{Mega-tree of mammals from PHYLACINE.}
\format{A phylogeny with class "phylo". It is also a list. Compare with a normal phylo object, it has another data frame \code{tree_mammal$genus_family_root}, which provides the root nodes information for every unique genus and family in the phylogeny. Such information can be used to insert new tips onto the phylogeny later.}
\source{
\url{https://github.com/MegaPast2Future/PHYLACINE_1.2/blob/master/Data/Phylogenies/Complete_phylogeny.nex}
}
\usage{
tree_mammal
}
\description{
This tree was the 288th \code{set.seed(123); sample(1:1000, 1)} of those 1000 trees generated by Faurby et al. (2018). Node labels were added here.
}
\references{
Faurby, S., Davis, M., Pedersen, R. Ø., Schowanek, S. D., Antonelli, A., & Svenning, J. C. (2018). PHYLACINE 1.2: The phylogenetic atlas of mammal macroecology. Ecology, 99(11), 2626-2626.
}
\keyword{datasets}
