% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tree_plant_GBOTB}
\alias{tree_plant_GBOTB}
\title{Mega-tree of plants based on Open Tree of Life}
\format{A phylogeny with class "phylo". It is also a list. Compare with a normal phylo object, it has another data frame \code{tree_plant_GBOTB$genus_family_root}, which provides the root nodes information for every unique genus and family in the phylogeny. Such information can be used to insert new tips onto the phylogeny later.}
\source{
\url{https://github.com/jinyizju/V.PhyloMaker/tree/master/data}
}
\usage{
tree_plant_GBOTB
}
\description{
This tree was based on Brown and Smith (2018), which in turn was based on the Open Tree of Life.
It was copied from \code{V.PhyloMaker::GBOTB.extended}. After then, node labels were added for empty ones.
}
\keyword{datasets}
