% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{ValidateDriver}
\alias{ValidateDriver}
\title{Validate Driver}
\usage{
ValidateDriver(Driver = "driver.txt")
}
\arguments{
\item{Driver}{Data frame or path of driver to be validated.}
}
\value{
The number of issues detected. Details are printed to console.
}
\description{
Validate Driver
}
\examples{
bdr<-BuildDriver()
bdr[matrix(
  sample(
    x = c(TRUE,FALSE),
    size = nrow(bdr)*ncol(bdr),
    prob = c(0.01,0.99),
    replace = TRUE
  ),
  nrow = nrow(bdr),
  ncol = ncol(bdr)
)]<- NA
ValidateDriver(Driver = bdr)

}
