% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{ValidateParameters}
\alias{ValidateParameters}
\title{Validate Parameters and attempt repairs}
\usage{
ValidateParameters(parameters = "param.p")
}
\arguments{
\item{parameters}{Data.frame or file path for TREES parameters in need of validation.}
}
\value{
data frame of repaired parameters.
}
\description{
Validate Parameters and attempt repairs
}
\examples{
demo_p<-ExampleParameters()
demo_good_p_valid<-ValidateParameters(demo_p)
demo_p_bad<-demo_p
demo_p_bad[matrix(
  sample(x = c(TRUE,FALSE),
         size = nrow(demo_p_bad)*ncol(demo_p_bad),
         prob = c(0.01,0.99),
         replace = TRUE
  ),
  nrow = nrow(demo_p_bad),
  ncol = ncol(demo_p_bad)
)]<-NA
demo_p_bad<-demo_p_bad[sample(1:174,170,replace = FALSE),]
demo_bad_p_valid<-ValidateParameters(demo_p_bad)
}
