% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{Var_unit}
\alias{Var_unit}
\title{Unit expression to help with plotting}
\usage{
Var_unit(variable_name)
}
\arguments{
\item{variable_name}{The name of the variable you want the expression for. Currently only supports sim_out variables.}
}
\value{
object of class expression for labels in ggplot axis.
}
\description{
Unit expression to help with plotting
}
