% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{Visual_driver}
\alias{Visual_driver}
\title{Driver parameter visualizer.}
\usage{
Visual_driver(driver, dailys = FALSE)
}
\arguments{
\item{driver}{Data.frame or file path for driver. Must have column names matching Driver_inputs()}

\item{dailys}{TRUE or FALSE aggregate data to daily values?}
}
\value{
ggplot object facet wrapped to view each of the driver inputs.
}
\description{
Driver parameter visualizer.
}
\examples{
Visual_driver(driver = BuildDriver())
Visual_driver(driver = BuildDriver(),dailys = TRUE)

}
