% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{Write_TREES_files}
\alias{Write_TREES_files}
\title{Write out original style TREES files for sharing changes}
\usage{
Write_TREES_files(
  destination = getwd(),
  prefix = "demo",
  driver = NA,
  parameter = NA,
  rss = NA
)
}
\arguments{
\item{destination}{Path to store files to.}

\item{prefix}{Character string to prepend all files with.}

\item{driver}{Valid driver data.frame.}

\item{parameter}{Valid parameter data.frame.}

\item{rss}{valid root_shoot_soil list.}
}
\value{
Returns a vector of the locations where the files were written.
}
\description{
Write out original style TREES files for sharing changes
}
\examples{
#'
#Write_TREES_files(
  #destination="inst",#"path/to/destination",#use whole path if outside working dir
  #prefix="demo",
  #driver=BuildDriver(),
  #parameter=ExampleParameters(),
  #rss=ExampleRSS()
#)
}
