% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_root_info}
\alias{add_root_info}
\title{Add genus and family basal/root node information to a phylogeny}
\usage{
add_root_info(tree, classification)
}
\arguments{
\item{tree}{A phylogeny with class "phylo".}

\item{classification}{A data frame of 2 columns: genus, family. It should include
all genus the tips of the tree belong to.}
}
\value{
A phylogeny with basal nodes information attached.
}
\description{
Based on the classification of tips, find where is the basal and root node for
each genus and each family. Such information can be later used to graft new
tips onto the phylogeny. This function can be used to process a user provided
tree.
}
