% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_root_info}
\alias{add_root_info}
\title{Add genus and family basal/root node information to a phylogeny}
\usage{
add_root_info(
  tree,
  classification,
  process_all_tips = TRUE,
  genus_list = NULL,
  family_list = NULL,
  show_warning = FALSE
)
}
\arguments{
\item{tree}{A phylogeny with class "phylo".}

\item{classification}{A data frame of 2 columns: genus, family. It should include
all genus the tips of the tree belong to.}

\item{process_all_tips}{Whether to find basal nodes for all tips? Default is \code{TRUE}.}

\item{genus_list}{An optinoal subset list of genus to find root information.}

\item{family_list}{An optinoal subset list of family to find root information.
This should be for species that do not have co-genus in the tree.}

\item{show_warning}{Whether to print warning information about non-monophyletic clades or not.}
}
\value{
A phylogeny with basal nodes information attached.
}
\description{
Based on the classification of tips, find where is the basal and root node for
each genus and each family. Such information can be later used to graft new
tips onto the phylogeny. This function can be used to process a user provided
tree.
}
