% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{classifications}
\alias{classifications}
\title{Classifications of species}
\format{A data frame with three columns: genus, family, and taxon (\code{plant}, \code{fish}).}
\usage{
classifications
}
\description{
Genus and family information of different groups of taxon.
\itemize{
\item Plant classification information was based on The Plant List. It was generated
based on \code{V.PhyloMaker::nodes.info.1} with two additions: genus Malaisia (Euphorbiaceae)
and Lithraea (Anacardiaceae). Three are 10,606 genus in this file.
\url{https://github.com/jinyizju/V.PhyloMaker/tree/master/data}
\item Fish classification information was based on FishBase. They are 4,825 genus in this file.
\url{https://fishtreeoflife.org/downloads/PFC_taxonomy.csv.xz}
}
}
\keyword{datasets}
