% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{classifications}
\alias{classifications}
\title{Classifications of species}
\format{
A data frame with three columns: genus, family, and taxon (\code{plant}, \code{fish}, \code{bird}, \code{mammal}, \code{amphibian}, \code{reptile}, \code{shark_ray}, \code{bee}, \code{butterfly}).
}
\usage{
classifications
}
\description{
Genus and family information of different groups of taxon.
\itemize{
\item Plant classification information. Its sources include:
+ based on \href{https://github.com/jinyizju/V.PhyloMaker/tree/master/data}{\code{V.PhyloMaker::nodes.info.1}}
+ based on The Plant List
+ \href{https://github.com/traitecoevo/taxonlookup}{taxonlookup}
+ \href{http://plantsoftheworldonline.org/}{Plants of the World online}
\item Fish classification information was based on FishBase. There are 4,825 genus in this file.
\url{https://fishtreeoflife.org/downloads/PFC_taxonomy.csv.xz}
\item Bee classification information was from \href{http://beetreeoflife.org/nomenclature/}{Bee Tree of Life}. Note that we used 'Subfamily' in their nomenclature file as "family" here. If a genus' Subfamily is missing, we used its Family.
\item Bird classification information was based on BirdLife, which resulted in 2,391 genus.
\url{http://datazone.birdlife.org/species/taxonomy} However, based on the
\href{https://data.vertlife.org/birdtree/BLIOCPhyloMasterTax.csv}{taxonomy file}
of the Jetz et al. 2012 phylogeny, there are additional 117 genus that are not in the file
of BirdLife. Both are combined here, which leads to 2,508 genus.
\item Mammal classification information was based on PHYLACINE, which has 1,400 genus.
\url{https://github.com/MegaPast2Future/PHYLACINE_1.2/blob/master/Data/Taxonomy/Synonymy_table_valid_species_only.csv} Additional
genus from Vertlife were added too. For the same genus from both PHYLACINE and Vertlife that have different family information,
I used the family from Vertlife as I found that they are mostly more accurate.
\item Amphibian classification information was from \href{https://data.vertlife.org/amphibiantree/download/amph_shl_new_Classification.csv}{VertLife}.
\item Reptile classification information was largely from \href{https://en.wikipedia.org/wiki/List_of_reptile_genera}{wikipedia}.
\item Shark and Ray classification information was largely from NCBI.
\item Butterfly classification information was from Kawahara et al. 2023, using the tip labels of their phylogeny.
}
}
\keyword{datasets}
