% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tree.R
\name{get_tree}
\alias{get_tree}
\title{Derive a phylogeny from a mega-tree}
\usage{
get_tree(sp_list, tree, taxon, scenario = c("S3", "S2", "S1"),
  tree_by_user = FALSE)
}
\arguments{
\item{sp_list}{A data frame with at least three columns: species, genus, family. Species column
holds the species for which we want to have a phylogeny.}

\item{tree}{A mega-tree with class \code{phylo}.}

\item{taxon}{The taxon of species in the \code{sp_list}.}

\item{scenario}{How to insert a species into the mega-tree?
\itemize{
\item If \code{scenario = "S1"}, a species is attached to the basal node of the same genus or the same family
if the mega-tree does not have any species of this genus.
\item If \code{scenario = "S2"}, a species is attached to a randomly selected node that is at or below the
basal node of the same genus of the same family if the mega-tree does not have any species in this genus.
Because of the random sampling involved, you may want to run several times to get a collection of
derived phylogenies.
\item If \code{scenario = "S3"}, a species is attached to the basal node of the same genus if the mega-tree has species
in the same genus; otherwise, a species is inserted to a new node above the basal node of the family.
If the age of the basal node is less than 2/3 of the node above it (root node of the family), the new node will
be added so that its age will be 2/3 of the root node of the family. Otherwise, a new node will be inserted
into the middle point of the basal node and the root node of the family.
\item In all scenarioes, if there is only 1 species in the genus or family, a new node will be inserted to
the middle point of this only species' branch length and the new species will be attached to this new
node.
}}

\item{tree_by_user}{Is the mega-tree provided by user?}
}
\value{
A phylogeny for the species required, with class \code{phylo}.
}
\description{
For a list of species, generate a phylogeny from a provided mega-tree. If a species is
not in the mega-tree, it will be inserted to the mega-tree with three scenarioes.
}
