% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tree.R
\name{get_tree}
\alias{get_tree}
\title{Derive a phylogeny from a mega-tree}
\usage{
get_tree(
  sp_list,
  tree,
  taxon,
  scenario = c("S1", "S2", "S3"),
  show_grafted = FALSE,
  tree_by_user = FALSE
)
}
\arguments{
\item{sp_list}{A data frame with at least three columns: species, genus, family. Species column
holds the species for which we want to have a phylogeny. It can also have two optional columns:
close_sp and close_genus. We can specify the cloest species/genus of the species based on
expert knowledgement. If specified, the new species will be grafted to that particular location.

It can also be a string vector if \code{taxon} is specified. Though it probably is a better idea
to prepare your data frame with \code{\link[=sp_list_df]{sp_list_df()}}.

In all cases, species names should be separated by \verb{_} between genus and species, i.e. genus_sp.}

\item{tree}{A mega-tree with class \code{phylo}. Optional if \code{taxon} is specified, in which case, a default
mega-phylogeny will be used.
\itemize{
\item For plant, the mega-tree is \link{tree_plant_otl}.
\item For fish, the mega-tree is \link{tree_fish}.
\item For bird, the mega-trees are \link{tree_bird_ericson} and \link{tree_bird_hackett}. \link{tree_bird_ericson} will be the
default if no \code{tree} is specified and \code{taxon} is \code{bird}.
\item For mammal, the mega-tree is \link{tree_mammal}.
}}

\item{taxon}{The taxon of species in the \code{sp_list}. Currently, can be \code{plant}, \code{fish}, \code{bird}, or \code{mammal}.}

\item{scenario}{How to insert a species into the mega-tree?
\itemize{
\item If \code{scenario = "S1"}, a species is attached to the basal node of the same genus or the same family
if the mega-tree does not have any species of this genus.
\item If \code{scenario = "S2"}, a species is attached to a randomly selected node that is at or below the
basal node of the same genus of the same family if the mega-tree does not have any species in this genus.
The probability of node been selected is proportional to its branch length.
Because of the random sampling involved, you may want to run several times to get a collection of
derived phylogenies.
\item If \code{scenario = "S3"}, a species is attached to the basal node of the same genus if the mega-tree has species
in the same genus; otherwise, a species is inserted to a new node above the basal node of the family.
If the age of the basal node is less than 2/3 of the node above it (root node of the family), the new node will
be added so that its age will be 2/3 of the root node of the family. Otherwise, a new node will be inserted
into the middle point of the basal node and the root node of the family. I probably won't use this scenario.
\item In all scenarioes, if there is only 1 species in the genus or family, a new node will be inserted to
the middle point of this only species' branch length and the new species will be attached to this new
node.
}}

\item{show_grafted}{Whether to indicate which species was grafted onto the mega-tree.
If \code{TRUE}, a \code{*} will be appended to the species name on the tip if it was grafted within
the same genus; \verb{**} will be appended if it was grafted within the same family.}

\item{tree_by_user}{Is the mega-tree provided by user?}
}
\value{
A phylogeny for the species required, with class \code{phylo}.
}
\description{
For a list of species, generate a phylogeny from a provided mega-tree. If a species is
not in the mega-tree, it will be grafted to the mega-tree with three scenarioes.
}
