% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tree.R
\name{get_tree}
\alias{get_tree}
\title{Get one or multiple trees from megatree(s)}
\usage{
get_tree(
  sp_list,
  tree,
  taxon,
  scenario = c("at_basal_node", "random_below_basal", "at_or_above_basal"),
  show_grafted = FALSE,
  tree_by_user = FALSE,
  mc_cores = 1
)
}
\arguments{
\item{sp_list}{A data frame with at least three columns: species, genus, family. Species column
holds the species for which we want to have a phylogeny. It can also have two optional columns:
close_sp and close_genus. We can specify the closest species/genus of the species based on
expert knowledge. If specified, the new species will be grafted to that particular location.

It can also be a string vector if \code{taxon} is specified. Though it probably is a better idea
to prepare your data frame with \code{\link[=sp_list_df]{sp_list_df()}}.

In all cases, species names should be separated by \verb{_} between genus and species, i.e. genus_sp.}

\item{tree}{A mega-tree with class \code{phylo} or a list of mega-trees with class \code{multiPhylo}.
Optional if \code{taxon} is specified, in which case, a default
mega-phylogeny will be used (see their own documentations for details).
\itemize{
\item For plant, the mega-tree is \link{tree_plant_otl}.
\item For fish, the mega-tree is \link{tree_fish}.
\item For bird, the mega-trees are \link{tree_bird_ericson} and \link{tree_bird_hackett}. \link{tree_bird_ericson} will be the
default if no \code{tree} is specified and \code{taxon} is \code{bird}.
\item For mammal, the mega-tree is \link{tree_mammal}.
}}

\item{taxon}{The taxon of species in the \code{sp_list}. Currently, can be \code{plant}, \code{fish}, \code{bird}, or \code{mammal}.}

\item{scenario}{How to insert a species into the mega-tree?
\itemize{
\item In all scenarioes, if there is only 1 species in the genus or family, a new node will be inserted to
the middle point of this only species' branch length and the new species will be attached to this new
node.
\item If \code{scenario = "at_basal_node"}, a species is attached to the basal node of the same genus or the same family
if the mega-tree does not have any species of this genus.
\item If \code{scenario = "random_below_basal"}, a species is attached to a randomly selected node that is at or below the
basal node of the same genus of the same family if the mega-tree does not have any species in this genus.
The probability of node been selected is proportional to its branch length.
Because of the random sampling involved, you may want to run several times to get a collection of
derived phylogenies.
\item If \code{scenario = "at_or_above_basal"}, a species is attached to the basal node of the same genus if the mega-tree has species
in the same genus; otherwise, a species is inserted to a new node above the basal node of the family.
If the age of the basal node is less than 2/3 of the node above it (root node of the family), the new node will
be added so that its age will be 2/3 of the root node of the family. Otherwise, a new node will be inserted
into the middle point of the basal node and the root node of the family. I probably won't use this scenario.
}}

\item{show_grafted}{Whether to indicate which species was grafted onto the mega-tree.
If \code{TRUE}, a \code{*} will be appended to the species name on the tip if it was grafted within
the same genus; \verb{**} will be appended if it was grafted within the same family.}

\item{tree_by_user}{Is the mega-tree provided by user? Default is \code{FALSE} but it will be automatically set to \code{TRUE} when the class of \code{tree} is
\code{multiPhylo} since we don't provide any such mega-trees here.}

\item{mc_cores}{Number of cores to parallel processing when \code{tree} is a list of large number of trees.}
}
\value{
A phylogeny for the species required, with class \code{phylo};
or a list of phylogenies with class \code{multiPhylo} depends on the input \code{tree}.
}
\description{
For some taxa groups, there are multiple posterior megatrees. It is a common
task to derive a phylogeny from each of these (or a random subset of) megatrees.
}
\details{
Derive a phylogeny from a mega-tree

For a list of species, generate a phylogeny or multiple phylogenies from a
provided mega-tree or mega-trees. If a species is
not in the mega-tree, it will be grafted to the mega-tree with three scenarios.
}
\examples{
test_sp = c("Serrasalmus_geryi", "Careproctus_reinhardti", "Gobiomorphus_coxii", 
"Periophthalmus_barbarus", "Prognichthys_glaphyrae", "Barathronus_bicolor", 
"Knipowitschia_croatica", "Rhamphochromis_lucius", "Neolissochilus_tweediei", 
"Haplochromis_nyanzae", "Astronesthes_micropogon", "Sanopus_reticulatus")
test_tree = get_tree(sp_list = test_sp,
                     tree = tree_fish, # either 
                     taxon = "fish", # or
                     scenario = "at_basal_node",
                     show_grafted = TRUE)
}
