% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtrees-package.R
\docType{package}
\name{rtrees-package}
\alias{rtrees-package}
\alias{_PACKAGE}
\alias{rtrees}
\title{Fetching phylogenies from mega-trees}
\description{
To facilitate generating phylogenies from synthesis trees.
}
\author{
Daijiang Li \href{mailto:daijianglee@gmail.com}{daijianglee@gmail.com}
}
\keyword{internal}
